fix: handle fix api warehouse

+++ src/main/java/vn/edu/iuh/fit/smartwarehousebe/dtos/requests/warehouse/UpdateWarehouseRequest.java
@@ -5,5 +5,5 @@ import java.io.Serializable;
 /**
  * DTO for {@link vn.edu.iuh.fit.smartwarehousebe.models.Warehouse}
  */
-public record UpdateWarehouseRequest(Integer id, String address, String code, String name, Integer managerId) implements Serializable {
+public record UpdateWarehouseRequest(Long id, String address, String code, String name, Long managerId) implements Serializable {
 }
